FI-3598: Add title to auth inputs and consolidate input modes (#591)
* add card to auth input mode

* move auth files to auth folder

* consolidate auth and access
