+++ client/src/components/InputsModal/Auth/AuthSettings.ts
